pyproject.toml
[project]
name = "baywatchers"
version = "0.1.0"
description = "Bot de seguimiento de precios con Telegram"
authors = [
    {name = "BaywatchZEfron",email = "alandor324@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4 (==4.13.4)",
    "python-decouple (==3.8)",
    "python-telegram-bot (==22.1)",
    "requests (==2.32.3)",
    "sqlalchemy (==2.0.41)",
    "apscheduler (==3.11.0)",
    "uvicorn (==0.34.2)",
    "fastapi (==0.115.12)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
